change: record rerun

+++ scripts/21_Upload.py
@@ -11,9 +11,9 @@ if __name__ == "__main__":
     # FTP settings
     FTP.encoding = "utf-8"
     ftp = FTP(
-        "xxx.xxx.xxx.xxx",
+        "192.168.10.230",
         "ftp_user",
-        passwd="xxxx"
+        passwd="FTPPassw0rd"
         )
     # -mp4ファイルの一覧を取得- #
     list_ext = ["mp4", "mp3"]
@@ -23,17 +23,22 @@ if __name__ == "__main__":
         # -アップロード処理- #
         for fil in files:
             # ディレクトリ作成処理
-            dir_parent01 = str(Path(dirPath).name)
-            dir_parent02 = str(fil.parents[0]).split("/")[-2]
-            dir_parent03 = str(fil.parents[0]).split("/")[-1]
-            dir_mkd = str(Path("FTP")/dir_parent01/dir_parent02/dir_parent03)
-            dir_mkd_parent = str(Path("FTP")/dir_parent01/dir_parent02)
-            if not dir_mkd in ftp.nlst(dir_mkd_parent):
-                ftp.mkd(dir_mkd + "/")  # 最後に/が必要
-            dir_mkd = dir_mkd + "/"
-            # upload処理
-            path_file = str(fil)
-            with open(path_file, "rb") as f:
-                ftp.storbinary("STOR " + dir_mkd + fil.name, f)
-            # アップロードしたファイルを削除
-            os.remove(path_file)
+            # 上位ディレクトリから順次作成
+            flag_mkdir = False
+            length_path = len(str(fil.parents[0]).split("/"))
+            dir_mkd = str(Path("FTP")/"record_files"/ext)
+            for i in range(length_path):
+                if flag_mkdir:
+                    dir_mkd_parent = dir_mkd
+                    dir_mkd = dir_mkd + "/" + fil.parents[length_path - (i+1)].name
+                    if not dir_mkd in ftp.nlst(dir_mkd_parent):
+                        ftp.mkd(dir_mkd + "/")  # 最後に/が必要
+                    # upload処理
+                    if i == (length_path - 1):
+                        dir_mkd = dir_mkd + "/"
+                        path_file = str(fil)
+                        with open(path_file, "rb") as f:
+                            ftp.storbinary("STOR " + dir_mkd + fil.name, f)
+                        # アップロードしたファイルを削除
+                        os.remove(path_file)
+                if(fil.parents[length_path - (i+1)].name == ext): flag_mkdir = True
